+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/StorageController.java
@@ -326,6 +326,10 @@ public final class StorageController {
             Log.w(LOGTAG, "Cannot set a tracking permission to VALUE_PROMPT, aborting.");
             return;
         }
+        if (type == GeckoSession.PermissionDelegate.PERMISSION_STORAGE_ACCESS) {
+            Log.w(LOGTAG, "Cannot set storage access permission via String API.");
+            return;
+        }
         final GeckoBundle msg = new GeckoBundle(5);
         msg.putString("uri", uri);
         msg.putString("contextId", createSafeSessionContextId(contextId));