+++ mobile/android/geckoview_example/src/main/java/org/mozilla/geckoview_example/GeckoViewActivity.java
@@ -1804,6 +1804,9 @@ public class GeckoViewActivity
                 case PERMISSION_MEDIA_KEY_SYSTEM_ACCESS:
                     resId = R.string.request_media_key_system_access;
                     break;
+                case PERMISSION_STORAGE_ACCESS:
+                    resId = R.string.request_storage_access;
+                    break;
                 default:
                     return GeckoResult.fromValue(ContentPermission.VALUE_DENY);
             }