+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -13,6 +13,13 @@ exclude: true
 
 ⚠️  breaking change and deprecation notices
 
+## v92
+- Added [`GeckoSession.PermissionDelegate.PERMISSION_STORAGE_ACCESS`][92.1] to
+  control the allowing of third-party frames to access first-party cookies and
+  storage. ([bug 1543720]({{bugzilla}}1543720))
+
+[92.1]: {{javadoc_uri}}/GeckoSession.PermissionDelegate.html#PERMISSION_STORAGE_ACCESS
+
 ## v91
 - Extended [`Autocomplete`][78.7] API to support addresses.
   ([bug 1699794]({{bugzilla}}1699794)).
@@ -1022,4 +1029,4 @@ to allow adding gecko profiler markers.
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport-android.content.Context-android.os.Bundle-java.lang.String-
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: 05c7e21b7ea40cd9978b319a9acfac1b256e700f
+[api-version]: 6cc809570a2e6e41a82ba612eee0f3059aecb14d